make the project go gettable
