+++ Dockerfile
@@ -1,10 +1,10 @@
 FROM golang:1.8.3 as build
-COPY . /build
-WORKDIR /build
-RUN CGO_ENABLED=0 make build
-ENTRYPOINT ["bash"]
+COPY . /go/src/github.com/coyove/goflyway
+WORKDIR /go/src/github.com/coyove/goflyway
+RUN go install
 
 FROM scratch
-COPY --from=build /build/build/goflyway /goflyway
+COPY --from=build /go/bin/goflyway
+COPY goflyway.conf
 EXPOSE 8102 8100 8101
 ENTRYPOINT ["/goflyway"]
