+++ pkg/proxy/counter_darwin.go
