+++ pkg/proxy/counter_windows.go
